Refactor into sub packages
The purpose is to split the file into several smaller files.
I didn't touch the logic.
