+++ .gitignore
@@ -1,2 +1,3 @@
 /service.key
 /.vscode
+.idea
