+++ internal/config/config.go
@@ -1,0 +1,9 @@
+package config
+
+// Config represents Application global config.
+type Config struct {
+	OpenAPIURL string
+	DateFormat string
+	ServiceKey string
+	Port       string
+}
